+++ examples/sources/ImageSource.js
@@ -68,10 +68,10 @@ class ImageSource extends Component {
     }
 
     onRequestClose(e) {
-        const { onComplete } = this.props;
+        const { onClose } = this.props;
         e.preventDefault();
 
-        onComplete();
+        onClose();
     }
 
     onAfterOpen() {
@@ -119,6 +119,7 @@ class ImageSource extends Component {
 ImageSource.propTypes = {
     editorState: PropTypes.object.isRequired,
     onComplete: PropTypes.func.isRequired,
+    onClose: PropTypes.func.isRequired,
     entityType: PropTypes.object.isRequired,
     entityKey: PropTypes.string,
     entity: PropTypes.object,